Add description as camera attr. in delete methods and add warning in deleteall.

+++ src/main/java/org/eam/code/vmixapp/controller/OPController.java
@@ -454,6 +454,9 @@ public class OPController implements Initializable {
 
     @FXML
     void deleteAllCams() {
+        if(!Alarm.confirmationDelete("All non attached cameras.")) {
+            return;
+        }
         StringBuilder deleteResponse = myCameraService.deleteAllCams();
         if(!deleteResponse.isEmpty()) {
             Alarm.showError(deleteResponse.toString());
